nlp_mcrs.py and nlp_challenges.py
Introducing the nlp_mcrs.py and nlp_challenges.py scripts

nlp_mcrs.py: The Multi-Challenge Robot Script is designed to eliminate the need for multiple scripts for each challenge. The current version supports the Receptionist challenge, making the nlp_receptionist.py script redundant. The next step is to integrate support for other challenges, such as the Restaurant challenge.

Why do this? Currently, the nlp_gspr.py script provides a lot of useful information, but its output must be filtered to extract the required details. As a result, the nlp_receptionist.py script has been preferred so far for the Receptionist challenge, as it produces concise, easily processable output with the necessary information. The same concept is planned to be applied to other challenges. To achieve this, it is easier and more advantageous to create a single script that combines the well-structured NLP process from nlp_gspr.py with flexible support for different challenges.

The code for various challenges is defined in nlp_challenges.py. The required code for each challenge is called by a switch function, which is based on the intent extracted from the audio processing. For example, if the intent is "Receptionist," the switch function calls the receptionist function from the Receptionist class defined in nlp_challenges.py.