Restaurant challenge and changes to nluInternal in nlp_mcrs.py
Established the Restaurant class in nlp_challenges.py.  Similar functionality to Receptionist, but returns food and drinks instead. Also returns the number each food was ordered. Example output:

"<ORDER>, [['steaks', 'two'], ['fries', 'one']], [['tea', 'one']]"

For this to work, changes needed to be made to the nluInternal function in nlp_mcrs.py. pAdj now contains a list with all the entities, that are processed before appenind to the entitities list. This is needed so that more than one item is contained in the output of nluInternal. E.g. a food Item and a drink Item.