getData no longer class bound. Changes to receptionist
Instead of having a getData function for each class in nlp_challenges, we now have a getData function that works for each class.

Furthermore, changes have been made to receptionist, since "names", "drinks" and "foods", that are produced by getData, now also contain the amount of each instance. receptionist had to be adjusted to these changes