nlp_mcrs can use audio files now. nlp_restaurant.py. test.py
nlp_mcrs.py can now be run with a -a flag, that states that we want to use an .wav file as input. The file-path is currently directly statet in the script, will be changed later

nlp_restaurant.py is a backup script, in case nlp_mcrs.py fails.

test.py will be used to test the nlp scripts. Is still work in progress