switch function moved. Strings are translated to numbers now. Affirm and Deny implemented
Moved the switch function from nlp_mcrs.py to challenges.py. This is done to prepare for a merge of mcrs and gpsr. The idea is to only keep gpsr and in the long run have it ready to be rolled out as a package. For this reason, RoboCup and challenge specific code is now to be collected in nlp_challanges.py

Implemented Affirm and Deny. Mainly had to update the rasa model but also updated the filter, that skips any trasnciped sentence witch no entities, since affirm and deny usually contain none.

Fixed an issue where the numberAttribute for entities sometimes shows a string and at other time uses intengers. Strings are now converted to integers before publishing to ease the workload of Planning.