New Branch
Created a new branch to temporarily work on. This is due to the need of making significant changes to the nlp_mcrs.py script, in order to test the listen2queue method while using the backpack microphone (curently the method is never called when using it).
The "alternative" branch will not be used for this, since it currently holds changes for te planned integration of voice detection.