Create transcription_tests.py
Created a new test script, that tests the transcription results as well as the accuracy of the recognized intent and entities. Furthermore the current (intended) usecase compares a base transcription to a context enhanced transcription result.

Lastly some of the functionality of previous tests scripts has been reused and optimized, thus rendering them obsolete. They will be removed.